Remove aws folder and move cloudprovider specific tests
