--- tests/aws/terraform/ecr.tf
@@ -1,4 +1,0 @@
-resource "aws_ecr_repository" "testrepo" {
-  name                 = "flux-test-repo-${random_pet.suffix.id}"
-  image_tag_mutability = "MUTABLE"
-}
