--- tests/aws/terraform/main.tf
@@ -1,7 +1,0 @@
-provider "aws" {}
-
-resource "random_pet" "suffix" {}
-
-locals {
-    name = "flux-test-${random_pet.suffix.id}"
-}
